Moves package.json from root to install directory and updated arches-project script to copy package.json to a project upon creation. re #4841
