--- .yarnrc
@@ -1,2 +1,0 @@
---install.modules-folder "./arches/app/media/packages"
---add.modules-folder "./arches/app/media/packages"
