+++ arches/install/.yarnrc
@@ -1,0 +1,2 @@
+--install.modules-folder "../app/media/packages"
+--add.modules-folder "../app/media/packages"
