--- arches/install/arches-templates/project_name/package.json
@@ -1,61 +1,0 @@
-{
-    "name": "arches",
-    "repository": {
-        "url": "git@github.com:archesproject/arches.git",
-        "type": "git"
-    },
-    "dependencies": {
-        "@mapbox/geojson-extent": "0.3.2",
-        "@mapbox/mapbox-gl-draw": "1.0.8",
-        "@turf/turf": "4.4.0",
-        "backbone": "1.3.3",
-        "blueimp-gallery": "2.15.2",
-        "bootstrap": "3.3.2",
-        "bootstrap-colorpicker": "^2.3.6",
-        "chosen-js": "1.8.2",
-        "ckeditor": "~4.6.2",
-        "codemirror": "^5.24.0",
-        "core-js": "^2.5.7",
-        "d3": "3.5.17",
-        "datatables.net": "~1.10.12",
-        "datatables.net-bs": "~1.10.12",
-        "datatables.net-buttons": "^1.5.4",
-        "datatables.net-buttons-bs": "^1.5.4",
-        "datatables.net-responsive": "~2.1.0",
-        "datatables.net-responsive-bs": "~2.1.0",
-        "dom4": "^2.0.1",
-        "dropzone": "3.8.4",
-        "eonasdan-bootstrap-datetimepicker": "~4.17.43",
-        "flexslider": "~2.6.3",
-        "font-awesome": "~4.6.3",
-        "ionicons": "~2.0.1",
-        "jqtree": "1.3.4",
-        "jquery": "3.3.1",
-        "jquery-migrate": "3.0.1",
-        "jquery-validation": "1.17.0",
-        "jquery.easing": "^1.4.1",
-        "jqueryui": "1.11.1",
-        "js-cookie": "2.1.1",
-        "knockout": "3.4.0",
-        "knockout-mapping": "~2.6.0",
-        "knockstrap": "1.3.2",
-        "latlon-geohash": "1.1.0",
-        "leaflet": "1.0.3",
-        "leaflet-draw": "^0.4.9",
-        "leaflet-iiif": "^1.1.1",
-        "lt-themify-icons": "^1.1.0",
-        "mapbox-gl": "0.48.0",
-        "mathjs": "^3.11.4",
-        "metismenu": "^2.7.2",
-        "moment": ">=2.10.5",
-        "nouislider": "11.0.3",
-        "nvd3": "^1.8.5",
-        "proj4": "^2.3.15",
-        "requirejs": "~2.3.2",
-        "requirejs-plugins": "1.0.2",
-        "requirejs-text": "2.0.12",
-        "select2": "3.5.1",
-        "underscore": "1.9.1",
-        "uuidjs": "^3.3.0"
-    }
-}
